Assistant checkpoint: Fix tries counter and standardize container widths
Assistant generated file changes:
- static/js/main.js: Fix tries counter and update slider event
- static/css/custom.css: Update container widths and form styling

---

User prompt:

(2) the slider containment worked! But now it seems like the # of tries doesn't update as you move the slider
(3) this worked! Nice!
(1) Nice work! The width of the Sends div is great now! Can you make sure Standings, Sessions and Stats
(New) Can you replace the color of the div that contains the input forms to match the background? And can you remove the border on this div?